Fix power-on-via-advertise wiring for disconnected TV wake
The disconnected-state trigger for the manufacturer-data wake sequence
lived in press(), which the Power button/action never calls (they go
through pressSpecial()). Move it there, keyed on SPECIAL_POWER.

Also: powerAdvertStart() now actually starts the Data1/Data2 toggle
cycle; fix the powerAdvertStop() syntax error and restore-to-empty
manufacturer data; widen _power_advert_delay from uint8_t (overflows
at 1000) to uint32_t; and give the advert timeout chain its own timer
name so it can't be cancelled by an unrelated key-release timer.

+++ components/ble_mi_remote/ble_mi_remote.h
@@ -114,7 +114,7 @@ namespace esphome {
 				uint8_t				batteryLevel;
 				bool				_connected = false;
 				uint32_t			_delay_ms = 7;
-				uint8_t				_power_advert_delay = 1000;
+				uint32_t				_power_advert_delay = 1000;
 				uint8_t				_power_advert_cycle = 0;
 
 				uint16_t sid		= 0x01;
